+++ examples/read-all.js
@@ -1,14 +1,17 @@
 'use strict'
 
-const PromiseReadable = require('../lib/promise-readable')
 const fs = require('fs')
 
-async function main () {
-  const rstream = new PromiseReadable(fs.createReadStream(process.argv[2] || '/etc/hosts', {
-    highWaterMark: Number(process.argv[3]) || 1024
-  }))
+const {PromiseReadable} = require('../lib/promise-readable')
+
+async function main() {
+  const rstream = new PromiseReadable(
+    fs.createReadStream(process.argv[2] || __filename, {
+      highWaterMark: Number(process.argv[3]) || 1024,
+    }),
+  )
   const data = await rstream.readAll()
-  if (data != null) {
+  if (data !== undefined) {
     console.log(`Read ${data.length} bytes in total`)
     rstream.destroy()
   }