+++ examples/read-by-chunks.js
@@ -1,16 +1,19 @@
 'use strict'
 
-const PromiseReadable = require('../lib/promise-readable')
 const fs = require('fs')
 
-async function main () {
-  const rstream = new PromiseReadable(fs.createReadStream(process.argv[2] || '/etc/hosts', {
-    highWaterMark: Number(process.argv[3]) || 1024
-  }))
+const {PromiseReadable} = require('../lib/promise-readable')
+
+async function main() {
+  const rstream = new PromiseReadable(
+    fs.createReadStream(process.argv[2] || __filename, {
+      highWaterMark: Number(process.argv[3]) || 1024,
+    }),
+  )
 
   let total = 0
 
-  for (let chunk; (chunk = await rstream.read());) {
+  for (let chunk; (chunk = await rstream.read()); ) {
     console.log(`Read ${chunk.length} bytes chunk`)
     total += chunk.length
   }